% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizedata.R
\name{summarize_data}
\alias{summarize_data}
\title{Title: Summarize the data table}
\usage{
summarize_data(data, ..., summ_vars, na.rm = FALSE)
}
\arguments{
\item{data}{a data set in the data frame format or tibble format}

\item{...}{ellipse, indicating one or multiple grouping variables in \code{data}, has to be a categorical variable existed in the dataset to group the population}

\item{summ_vars}{a summarizing variable must be numeric variable and existed in the dataset}

\item{na.rm}{logical variable to indicate if missing value should be removed or not. default is FALSE(not remove)}
}
\value{
A summary tibble with calculated  mean, median and standard deviation of the indicated variables.
}
\description{
: This function is to summarize the numeric variable in a dataset when given one or more grouping conditions. It will generate a summary tibble to show the calculated median(median_v), mean (mean_v) and standard deviation (sd_v_) of the numeric variable under the indicated grouping conditions. There is an option for user to ignore missing value with na.rm arguement.
}
\details{
This function is utilizing \code{group_by}and \code{summarize()}functions in \code{dplyr} package
}
\examples{
data("penguins",package="palmerpenguins")
summarize_data(penguins,species,island,summ_vars=bill_length_mm,na.rm=TRUE)
summarize_data(penguins,sex,island,summ_vars=bill_depth_mm,na.rm=TRUE)
}
